+++ e2e/pkg/config/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package e2ecfg
+package config
 
 import (
 	"flag"
