+++ e2e/pkg/tests/policy/policy.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package calico
+package policy
 
 import (
 	"context"
